More character escapes + row bug fix.
We were previously keeping track of the starting column of a string or
character but not the starting row. The thought was that the all valid
strings and chars start and end on the same line. This turned out not to
be true if the Xi source code has a literal \f in it. This is a bit
problematic having the Xi newline and the Java newline be different, so
we'll have to investigate further.