Made things build.
The old Parser.cup file was not building. This made it hard to build
other things. I commented out everything in the Parser to make it build
only to discover that it didn't build! Turns out we had to remove our
other cup jars and use only Meyers' CUP jar.