Added Generic Type to AST.
Our old AST looked something like this

    type expr = Int of int
              | Bool of bool
              | ...

We didn't have row,col information in the AST which we needed for error
reporting during semantic analysis. We could add in a row, col in the
superclass of everythign and make it an abstract class. Unfortunately,
this doesn't play well with [1].

Instead, we can insert a generic argument to the type. For now, we pass
in row and column information, but later, we might throw in type
information! It's a double whammy.

    type 'a expr = Int of 'a * int
                 | Bool of 'a * bool
                 | ...

[1]: http://stackoverflow.com/questions/29740078/how-to-call-super-constructor-in-lombok